Fix broken user specs

+++ app/controllers/users_controller.rb
@@ -40,10 +40,14 @@ class UsersController < ApplicationController
   end
   
   def destroy
-    user = User.find_by(username: params[:id])
-    user.destroy
-    flash[:success] = "User #{params[:id]} has been deleted."
-    redirect_to users_url
+    user = User.find_by(username: params[:id], admin: false)
+    if user
+      user.destroy
+      flash[:success] = "User #{params[:id]} has been deleted."
+    else
+      flash[:warning] = "There was a problem deleting user #{params[:id]}"
+    end
+    redirect_to users_path
   end
 
   def confirm_email
